moved files out of products and created a ProductService
