+++ MyProject/src/Domain/Interfaces/IProductRepository.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using MyProject.Models;
+using MyProject.Models.Dto;
+using MyProject.src.Models;
+
+namespace MyProject.Interfaces
+{
+    public interface IProductRepository
+    {
+        Task<Product> CreateProductAsync(ProductCreateDto productCreateDto);
+        Task<Product> GetProductAsync(int id);
+    }
+}
