+++ MyProject/src/Domain/Interfaces/IProductService.cs
@@ -4,10 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using MyProject.Models;
 using MyProject.Models.Dto;
+using MyProject.src.Models;
 
 namespace MyProject.Interfaces
 {
-    public interface IProductRepository
+    public interface IProductService
     {
         Task<Product> CreateProductAsync(ProductCreateDto productCreateDto);
         Task<Product> GetProductAsync(int id);
