+++ MyProject/src/Domain/Models/Product.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MyProject.Models
+namespace MyProject.src.Models
 {
     public class Product
     {
