+++ MyProject/src/Infrastructure/Repositories/ProductRepository.cs
@@ -1,0 +1,37 @@
+using MyProject.Data;
+using MyProject.Interfaces;
+using MyProject.Models.Dto;
+using MyProject.src.Models;
+
+namespace MyProject.Repositories
+{
+    public class ProductRepository : IProductRepository
+    {
+        private readonly MyDbContext _context;
+
+        public ProductRepository(MyDbContext context)
+        {
+            _context = context;
+        }
+
+        public async Task<Product> CreateProductAsync(ProductCreateDto productCreateDto)
+        {
+            var product = new Product
+            {
+                Name = productCreateDto.Name,
+                Price = productCreateDto.Price,
+                Description = productCreateDto.Description
+            };
+
+            await _context.Products.AddAsync(product);
+            await _context.SaveChangesAsync();
+
+            return product;
+        }
+
+        public async Task<Product> GetProductAsync(int id)
+        {
+            return await _context.Products.FindAsync(id);
+        }
+    }
+}
